feat: add confirmation pop-up for log out

+++ src/app/authorizing/components/authorizing/authorizing.component.ts
@@ -18,6 +18,8 @@ export class AuthorizingComponent implements OnInit {
   userName = this.userLogin
     ? this.filter.transform(this.usersService.getUsers(), this.userLogin)
     : '';
+  authFormVisible = false;
+  logOutConfirmVisible = false;
 
   constructor(
     public modalService: ModalServiceService,
@@ -42,26 +44,38 @@ export class AuthorizingComponent implements OnInit {
   showSingUpForm() {
     this.isUserNew = true;
     this.modalService.open();
+    this.authFormVisible = true;
   }
 
   showLoginForm() {
     this.isUserNew = false;
     this.modalService.open();
+    this.authFormVisible = true;
   }
 
-  closeSingUpForm() {
+  // closeSingUpForm() {
+  //   this.modalService.close();
+  // }
+
+  logout(data: boolean) {
+    if (data) {
+      this.userName = '';
+      this.toggleToken();
+      this.authorizeService.removeToken();
+    }
     this.modalService.close();
+    this.logOutConfirmVisible = false;
   }
 
-  logout() {
-    this.userName = '';
-    this.toggleToken();
-    this.authorizeService.removeToken();
+  openConfirm() {
+    this.modalService.open();
+    this.logOutConfirmVisible = true;
   }
 
   authorize(data: User | string) {
     if (typeof data === 'string') {
       this.modalService.close();
+      this.authFormVisible = false;
     } else {
       if (this.isUserNew) {
         this.authorizeService.signUp(data).subscribe((singUpRes) => {
@@ -73,12 +87,11 @@ export class AuthorizingComponent implements OnInit {
             .subscribe((res) => {
               if (res.token) {
                 this.authorizeService.setToken(res.token, data.login);
-                this.closeSingUpForm();
+                // this.closeSingUpForm();
+                this.modalService.close();
                 this.toggleToken();
                 this.usersService.saveUsers();
                 this.userLogin = data.login;
-
-                // console.log(this.usersComponent.users);
               }
             });
         });
@@ -91,7 +104,8 @@ export class AuthorizingComponent implements OnInit {
           .subscribe((res) => {
             if (res.token) {
               this.authorizeService.setToken(res.token, data.login);
-              this.closeSingUpForm();
+              // this.closeSingUpForm();
+              this.modalService.close();
               this.toggleToken();
               this.usersService.saveUsers();
               this.userLogin = data.login;
